# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Авторизация пользователя",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["auth"],
                "summary": "Вход пользователя",
                "parameters": [
                    {
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AuthorizeUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "token": { "type": "string" }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "type": "object",
                            "additionalProperties": { "type": "string" }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": { "type": "string" }
                        }
                    }
                }
            }
        },
        "/api/user/get": {
            "get": {
                "description": "Получает информацию о пользователе по токену",
                "tags": ["users"],
                "summary": "Информация о пользователе",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": { "type": "string" },
                                "phone_number": { "type": "string" },
                                "role": { "type": "string" },
                                "created_at": { "type": "string" },
                                "updated_at": { "type": "string" },
                                "first_name": { "type": "string" },
                                "last_name": { "type": "string" }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": { "type": "string" }
                        }
                    }
                }
            }
        },
        "/api/user/update": {
            "post": {
                "description": "Обновляет информацию о пользователе",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["users"],
                "summary": "Обновить пользователя",
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": { "type": "string" }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": { "type": "string" }
                        }
                    }
                }
            }
        },
        "/users": {
            "delete": {
                "description": "Удаляет пользователя",
                "tags": ["users"],
                "summary": "Удалить пользователя",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": { "type": "string" }
                        }
                    }
                }
            }
        },
        "/api/cars/{id}": {
            "get": {
                "description": "Возвращает данные автомобиля по ID",
                "tags": ["cars"],
                "summary": "Детали автомобиля",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Car"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": { "type": "string" }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": { "type": "string" }
                        }
                    }
                }
            }
        },
        "/api/cars/add": {
            "post": {
                "description": "Создаёт новый автомобиль",
                "tags": ["cars"],
                "summary": "Добавить автомобиль",
                "consumes": ["multipart/form-data"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "brand",
                        "in": "formData",
                        "type": "string"
                    },
                    {
                        "name": "model",
                        "in": "formData",
                        "type": "string"
                    },
                    {
                        "name": "year",
                        "in": "formData",
                        "type": "integer"
                    },
                    {
                        "name": "color",
                        "in": "formData",
                        "type": "string"
                    },
                    {
                        "name": "mileage",
                        "in": "formData",
                        "type": "integer"
                    },
                    {
                        "name": "price_per_day",
                        "in": "formData",
                        "type": "number"
                    },
                    {
                        "name": "status",
                        "in": "formData",
                        "type": "string"
                    },
                    {
                        "name": "location_id",
                        "in": "formData",
                        "type": "integer"
                    },
                    {
                        "name": "image",
                        "in": "formData",
                        "type": "file"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": { "type": "integer" },
                                "message": { "type": "string" },
                                "car": { "$ref": "#/definitions/models.Car" }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": { "type": "string" }
                        }
                    }
                }
            }
        },
        "/api/cars/brands": {
            "get": {
                "description": "Возвращает список всех марок автомобилей",
                "tags": ["cars"],
                "summary": "Список марок",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": { "type": "string" }
                        }
                    }
                }
            }
        },
        "/api/cars/models": {
            "get": {
                "description": "Возвращает список моделей для указанной марки",
                "tags": ["cars"],
                "summary": "Модели по марке",
                "parameters": [
                    {
                        "name": "brand",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "models": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": { "type": "string" }
                        }
                    }
                }
            }
        },
        "/api/cars/filter": {
            "post": {
                "description": "Фильтрует автомобили по заданным критериям",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["cars"],
                "summary": "Фильтр автомобилей",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SearchCarRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Car"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": { "type": "string" }
                        }
                    }
                }
            }
        },
        "/cars/{id}": {
            "delete": {
                "description": "Удаляет автомобиль по ID",
                "tags": ["cars"],
                "summary": "Удалить автомобиль",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": { "type": "string" }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AuthorizeUserRequest": {
            "type": "object",
            "properties": {
                "email": { "type": "string" },
                "password": { "type": "string" }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "id": { "type": "integer" },
                "email": { "type": "string" },
                "first_name": { "type": "string" },
                "last_name": { "type": "string" },
                "phone_number": { "type": "string" },
                "role": { "type": "string" },
                "created_at": { "type": "string" },
                "updated_at": { "type": "string" }
            }
        },
        "models.Car": {
            "type": "object",
            "properties": {
                "id": { "type": "integer" },
                "brand": { "type": "string" },
                "model": { "type": "string" },
                "year": { "type": "integer" },
                "color": { "type": "string" },
                "mileage": { "type": "integer" },
                "price_per_day": { "type": "number" },
                "status": { "type": "string" },
                "location_id": { "type": "integer" },
                "image_url": { "type": "string" },
                "created_at": { "type": "string" },
                "updated_at": { "type": "string" }
            }
        },
        "models.SearchCarRequest": {
            "type": "object",
            "properties": {
                "brand": { "type": "string" },
                "model": { "type": "string" },
                "year_from": { "type": "integer" },
                "year_to": { "type": "integer" },
                "min_price": { "type": "number" },
                "max_price": { "type": "number" },
                "start_date": { "type": "string" },
                "end_date": { "type": "string" }
            }
        },
...
    }
}